image semantic segmentation/P8_03_api/requirements.txt
azureml-core==1.21.0.post1
numpy==1.18.5
Flask==1.0.3
Pillow==8.1.2
requests==2.25.1
